+++ src/generate.c
@@ -14,14 +14,8 @@ unsigned short int print_secret(invocation_state_t *invocation,
 	authenticator_parameters_t *authenticator_params =
 	    build_authenticator_parameters_from_deserialized_cleartext_and_passphrase(
 	        cleartext, invocation->passphrase);
-	uint16_t device_vendor = cleartext->device_vendor;
-	uint16_t device_product = cleartext->device_product;
-	bool skip_device_checks = device_vendor == OBFUSCATED_DEVICE_SENTINEL &&
-	                          device_product == OBFUSCATED_DEVICE_SENTINEL;
 
 	// Clean up things we don't need anymore
-	free_cleartext(cleartext);
-	cleartext = NULL;
 	sodium_memzero(invocation->passphrase, strlen(invocation->passphrase));
 	free(invocation->device);
 	free(invocation->file);
@@ -31,14 +25,20 @@ unsigned short int print_secret(invocation_state_t *invocation,
 
 	for (size_t i = 0; i < devices_list->count; i++) {
 		const fido_dev_info_t *di = fido_dev_info_ptr(devices_list->list, i);
-		if (skip_device_checks ||
-		    (fido_dev_info_vendor(di) == device_vendor &&
-		     fido_dev_info_product(di) == device_product)) {
-			const char *authenticator_path = fido_dev_info_path(di);
-			fido_dev_t *authenticator = get_device(authenticator_path);
+		const char *authenticator_path = fido_dev_info_path(di);
+		fido_dev_t *authenticator = get_device(authenticator_path);
+		fido_cbor_info_t *device_info = get_device_info(authenticator);
+		if (device_aaguid_matches(cleartext, device_info)) {
+			if (!device_supports_hmac_secret(device_info)) {
+				free_device_info(device_info);
+				continue;
+			}
+
 			secret_t *secret = malloc(sizeof(secret_t));
 			int result = get_secret_consuming_authenticator_params(
 			    authenticator, authenticator_params, secret);
+			close_and_free_device_ignoring_errors(authenticator);
+			free_device_info(device_info);
 			if (result == FIDO_OK) {
 				for (int j = 0; j < secret->secret_size; j++) {
 					printf("%02x", secret->secret[j]);
@@ -49,17 +49,42 @@ unsigned short int print_secret(invocation_state_t *invocation,
 			warnx("%s at %s did not return a valid secret: %s (0x%x)",
 			      fido_dev_info_product_string(di), authenticator_path,
 			      fido_strerr(result), result);
-		} else {
-			printf("Found device with device_vendor=%x, device_product=%x\n",
-			       fido_dev_info_vendor(di), fido_dev_info_product(di));
 		}
 	}
 
+	free_cleartext(cleartext);
+	cleartext = NULL;
+
 	if (devices_list->count > 0) {
-		free_devices_list(devices_list);
 		return EXIT_NO_VALID_AUTHENTICATOR;
 	}
 
-	free_devices_list(devices_list);
 	return EXIT_NO_DEVICES;
+}
+
+bool device_aaguid_matches(deserialized_cleartext *cleartext,
+                           fido_cbor_info_t *device_info) {
+	if (cleartext->device_aaguid_size == 0) {
+		return true;
+	}
+
+	size_t this_device_aaguid_size = fido_cbor_info_aaguid_len(device_info);
+
+	if (this_device_aaguid_size != cleartext->device_aaguid_size) {
+		return false;
+	}
+
+	unsigned char *this_device_aaguid = malloc(cleartext->device_aaguid_size);
+	memcpy(this_device_aaguid, fido_cbor_info_aaguid_ptr(device_info),
+	       this_device_aaguid_size);
+
+	for (size_t i = 0; i < cleartext->device_aaguid_size; i++) {
+		if (cleartext->device_aaguid[i] != this_device_aaguid[i]) {
+			free(this_device_aaguid);
+			return false;
+		}
+	}
+
+	free(this_device_aaguid);
+	return true;
 }