+++ CHANGELOG.md
@@ -3,6 +3,14 @@ _These changes are on the branch `master`, but not yet in a versioned release._
 
 * _(None yet)_
 
-## Version 1.0.0
+## Version 0.2.0
+
+* Remove use of device vendor & product ID, replace with AAGUID
+* Fix double free bugs when the wrong authenticator device is used
+* Fix reference to algorithm as uint32 in serialzie.c
+* Add escaping for `-` in manpage.m4
+* Convert spaces to tabs in bash-completion.m4
+
+## Version 0.1.0
 
 * Initial release.